+++ app.py
@@ -2,7 +2,7 @@ import os
 from dotenv import load_dotenv
 import logging
 from flask import Flask, render_template, jsonify
-from database import engine
+from database import load_jobs_db
 from sqlalchemy import text
 
 load_dotenv()
@@ -13,27 +13,15 @@ app = Flask(__name__, template_folder=template_dir)
 
 logging.basicConfig(level=logging.INFO)
 
-def load_jobs_db():
-    try:
-        with engine.connect() as conn:
-            result = conn.execute(text("SELECT * FROM jobs"))
-            jobs_list = []
-            for dict_row in result.mappings():
-                jobs_list.append(dict(dict_row))
-            logging.info("Jobs loaded from DB")
-            return jobs_list
-    except Exception as e:
-        logging.info("Error occured while loading jobs from DB: ", e)
 
 @app.route("/")
 def home_page():
     try:
         jobs = load_jobs_db()
-        logging.info("Jobs info loaded ")
+        logging.info("Jobs info loaded on page")
         return render_template('home.html', jobs=jobs, company_name="Best Jobs")
     except Exception as e:
         logging.info("Error while jobs loading: ", e)
-
 
 
 @app.route("/api/jobs")
